Add an install script, update README
Summary:
Instead of the README providing instructions to install the
playground, we can just have a script for that.

The README only needs a short section pointing to the install
script; the rest of it can describe how to debug each layer
(the venv and pyre, the flask server, the service and gunicorn,
nginx).

Reviewed By: pradeep90

Differential Revision: D35752450

fbshipit-source-id: 2e9df45b705c50f611b9bc6521e154bc0ca90a2d