Unpin dependencies
Summary:
The playground requirements file looks like it was the output of running `pip freeze`.

This isn't going to work well unless we also pin the pyre version - pyre relies on
updated releases of various packages, and trying to pin them but not pyre will create
errors (I was hitting a click error when I tried to run pyre).

Instead, we should unpin everything so that a newly-spun-up-server gets the latest
version of all dependencies. We also want to specify a very minimal set of dependencies,
so that if (for example) pyre pins one dependency temporarily we will pick that up.

Reviewed By: grievejia

Differential Revision: D35750056

fbshipit-source-id: f7cfcbf956d9685ffa44c51fab6379fd1e0f53dc