Use 10 workers rather than 3 per instance, use sequential mode
Summary:
We're running pyre playground on a 10-cpu machine. Let's use one worker
(and therefore one pyre server) per CPU. We probably could do even more
assuming a request is IO-bound but that's probably enough (10 requests
per second, nginx and gunicorn will queue them up for us).

I also modify the application to run pyre in sequential mode,
which (a) is probably faster anyway for
a small change, and (b) ensures that
with many concurrent requests we aren't
spawning a ton of processes.

Reviewed By: shannonzhu

Differential Revision: D35750120

fbshipit-source-id: 625d08f2f443fcf005ca8d5ec7f526c77bb8d710