Use eager parsing in the server for now
Summary:
Because pysa model generators rely on `pyre query "dump_class_hierarchy()"` and on
the ability to *reverse* that class hierarchy, we need to ensure that in `pyre query` the
class hierarchy is complete at least for now.

As a result, we need to ensure that the pyre server can be started in a way where it
gets the full hierarchy. This quick fix sets a new Configuration.Analysis flag to force
eager loading in UnannotatedGlobalEnvironment (the flag gets used in AstEnvironment
to set `invalidated_modules`) and uses that flag in `serverCommand`.

We'll still get fast `pyre check` using lazy parsing, but this will be a significant regression
in cold start time until we can make the flag controllable from the outside.

Reviewed By: arthaud

Differential Revision: D35779965

fbshipit-source-id: 66b1cbfb182a358125beaae97e735162870d8585