Add keyboard shortcut
Summary:
Add an event listener so that the keyboard shortcuts
Shift+Enter, Ctrl+Enter, and Cmd+Enter all cause us
to submit the code.

This allows users to work on examples and check them in
an entirely keyboard-driven way, which really does make
a difference when iterating on a nontrivial snippet.

Reviewed By: dkgi

Differential Revision: D35783184

fbshipit-source-id: e3299559a37eaecfea26a8f0ddaa6628d54aa301