Fix error messaging on bad pyre config field types
Summary:
cc omreeg This is the root cause of the confusing error messages you ran into in D35789471. Feel free to patch this diff and stack your changes on top of if you want an immediate fix to the weird behavior - but it just changes the error message so in effect the bad config field input is triggering the right exception.

For some reason we were printing the entire configuration json object, which we were popping fields from already in the validation process, and creating a nonsense error when we did hit a field of the wrong type.

Reviewed By: grievejia

Differential Revision: D35799175

fbshipit-source-id: 58e5e948a520043b38b2ccd3fafc3715b3973cea