Remove integrity check in debug mode
Summary:
We have an old hook to perform a ClassHierarchy integrity
check when debug mode is enabled.

I don't think this is used much, but if it *were* used then
it would very likely fail when using the lazy parser, since
the class hierarchy for external types is no longer guaranteed
to be complete.

It's always possible to put this back in as part of debugging,
although I doubt it has any value today (it was probably added
at a time when we were concerned about this environment being
correct). To do it would require force-loading all modules, which
can be done when necessary.

Reviewed By: shannonzhu

Differential Revision: D35823758

fbshipit-source-id: 2b28da73b86b4c6b158ef1419c238c14f55877d6