Never compute the call graph of a function during the global fixpoint
Summary:
We should always compute call graphs prior to the global fixpoint, so that those call graphs are accounted for in the dependency graph, in order to properly re-analyze callers when a model of a function changes. Otherwise, we are introducing non-determinism in the analysis - the result will depend on which function was analyzed first.

To prevent all this, let's fix the call graph interface to disallow creating call graphs on the fly.

Differential Revision: D35802122

fbshipit-source-id: 3cb52530961fab049ca6a32bccc4b8ec394e9e79