Add a directory existence check in site package roots
Summary:
We discovered in a recent CI run of torchx that it's possible for
python to report site package directories that do not actually exist.

We didn't realize this could happen, but when it does it can lead to
ocaml trying to list a nonexistent directory, which crashes Pyre.

See
https://github.com/pytorch/torchx/pull/468
for a real-world example in github CI of this failure case

Reviewed By: grievejia

Differential Revision: D35832569

fbshipit-source-id: 6290ec1c10df527a3b34cc855b1cd0a33758f836