Remove `key_to_string`, rename `find_unsafe` to `get_exn`
Summary:
The `string_of_key` function is never used. I think it may have been
written to enable debugging of the hashing, but given that the hashing
isn't actually tied to this function in any way it's just dead weight
and could actually cause confusion if used for debugging.

The name `find_unsafe` is ambiguous - it sounds like it might be some
kind of parallelism-unsafe operation. This is not actually the case, it's
just a `get` that returns a `value` directly, throwing if none is found
rather than returning a `value option`. We should follow the ocaml
naming convention here.

Reviewed By: grievejia

Differential Revision: D35783491

fbshipit-source-id: ee43a5d048b8233a071b2d3b81a6cb3f8d97437c