Overlays [0]: factor out key- and value- functor logic
Summary:
In order to create shared-memory overlay tables, we're going to
need the ability to construct a Key type using KeyFunctor as
well as KeySet and KeyMap types without having to inherit all of
the global interface in `NoCache.S`.

Factoring out the key and value logic - which is the same regardless
of whether the table is global or has first-class semantics - will
let us do this without repeating ourselves.

Reviewed By: grievejia

Differential Revision: D35818074

fbshipit-source-id: 628838406d71883e575d142396c1ca56697e0f87