Move the taint analysis entrypoint under `interprocedural_analysis/taint`
Summary:
# Context
The `pyre analyze` command was first meant to be a generic command that would call into different static analyses (e.g, dead stores, taint analysis, etc). This is implemented with the `Interprodecural` framework that runs a global fixpoint, given a callgraph. Different analyses can be plugged in by calling `Interprocedural.AnalysisResult.Make (ResultType).Register (AnalysisType)`.

Unfortunately, with time, the `pyre analyze` command has been changed specifically for the taint analysis, to a point where even the entrypoint calls taint analysis specific functions (inlining decorators, computing class intervals, etc.). This is introducing technical debt, for instance, with parameters being threaded in more and more complicated ways.

The call graph construction of the interprocedural framework is also heavily specialized for the taint analysis (e.g, it adds taint-it-taint-out flags, and other things). This is basically to a point where it is unsound for other analyses, which was discovered when working on type inference.

# Proposed changes
I think we reached a point where we want to change the whole design of the interprocedural framework. The objective with the following diffs is to split the framework into small components that can be extended and re-used. For instance, the global fixpoint would be a proper functor instead of relying on registration.

# Trade-offs
The current interprocedural framework allows to run multiple global fixpoints in parallel (for instance, a dead store and a taint analysis). We would lose that feature with the proposed changes. In practice, this was never used.

Another downside would potentially be code duplication - most likely, for the entrypoint. I think this can be avoided with proper designs.

# In this diff
This diff starts the migration by moving the entrypoint of `pyre analyze` directly into the `interprocedural_analysis/taint` directly, since - as mentioned - it is heavily customized for pysa anyway.

Reviewed By: stroxler

Differential Revision: D35833510

fbshipit-source-id: 4fc38427e6ca301a6fe67a9324dfba0330601ea2