Factor out computation of per-statement defined locals.
Summary:
Problem: Uninitialized locals has two parts: (a) computing the defined locals at each statement and (b) checking if any of the variables read in a statement is undefined at that point.

I want to reuse (a). So factor it out for now.

Note that the test inputs are from the other uninitialized check tests. However, we get the defined-locals for each statement. This gives us a clear idea of what's defined where.

Reviewed By: grievejia

Differential Revision: D35766571

fbshipit-source-id: 207194427e3b096c08f629be3bde857b87dbe376