Don't strip leading underscores in Callable.Parameter.zip
Summary:
The function Callable.Parameter.zip is used to line up parameter
lists.

It has logic to strip underscores from argument names, which I think
might be some legacy nonsense from before we had a proper implementation
of positional-only arguments; the logic was incorrect because an
argument with only one leading underscore isn't actually positional,
and it is therefore incorrect (and unsound) to treat it as equivalent
to the same argument with no underscore.

Moreover, there's no need to strip double underscores because (I
confirmed in a scratch project) we create PositionalOnly parameters
for double underscores, and zip already handles these in a
manner that is not sensitive to the name

Reviewed By: grievejia

Differential Revision: D35857167

fbshipit-source-id: 5aa680960121b89c5398ffbb097b7b4202cebd61