Break up long behavioral_subtyping test
Summary:
If we make a change to the override typing logic, the test is
so long that it is very hard to pinpoint changes - this is particularly
bad because for some reason OUnit doesn't give us line numbers in the
test itself, which is a problem that we have in some test modules that
I've thus far been unable to diagnose.

I also updated a couple of misleading comments.

Reviewed By: grievejia

Differential Revision: D35857166

fbshipit-source-id: ae9aff8e31258265354959442f7775cc48cd8c0c