overlays [1] create SharedMemory.FirstClass
Summary:
The FirstClass.NoCache and FirstClass.WithCache modules serve two
purposes:
- They define an interface (`sig S`) exactly like the (global) NoCache and
  WithCache modules, except that
  - there's an additional opaque type `t` representing a table
  - there's a function `create` to produce a `t`.
  - every function takes an extra argument of type `t` representing
    the identity of the table we we are working with
  - the LocalChanges interface, which we never use, is omitted
- They also define an implementation in terms of integer key prefixes.
  - the global key type, defined in `TupleKey`, is `int * Key.t`.
  - the implementation requires that `create` always be run in the main
    process in a parallelism-free environment since it relies on bumping
    a ref. We can probably come up with a more thread-safe approach but
    in practice this is unlikely to matter in our code today.

I chose to put this module in `SharedMemory` rather than `Memory` because it
needs to replicate almost exactly the same API as the global shared memory
modules defined here, but can't actually reuse the same interface. I'll put
the overlayable table logic elsewhere.

NOTE: The `FromGlobal` functor isn't part of the public interface; it is needed because
due to module gymnasitcs: in order to avoid copy-pasting all of the NoCache logic
I need a way to parametrize it on `Global` so that it can be reused for both
`WithCache` and `NoCache`.

Reviewed By: grievejia

Differential Revision: D35818073

fbshipit-source-id: e31058a4613a59f952e03709938e1e0d804866f4