Use marks to highlight errors
Summary:
Make errors easier to see by highlighting them in the document.

There's a separate gutter feature that might be nicer, but I've failed to make it work for several hours now; I think we're probably better off with this than nothing and I may not figure out the gutter in time.

Reviewed By: pradeep90

Differential Revision: D35918593

fbshipit-source-id: c0f221bbb5c5a8c4e60836759947c6df9e51d665