add selectionRange to pyre's documentSymbol definition
Summary: selectionRange is a required field and is necessary for vscode to accept this symbol definition. A selectionRange is a subset of the 'range' field which should be illuminated when a symbol is picked, ie when the symbol is clicked on in the outline view of vscode, while range should cover the entire symbol definition. For now, we will make them equivalent.

Reviewed By: grievejia

Differential Revision: D35938910

fbshipit-source-id: 205ae88086e3818a038ecd1b17c19d4a9dd96ddb