Refactor the missing flow analysis [interprocedural 5/n]
Summary:
See D35983946 (https://github.com/facebook/pyre-check/commit/4a29863f283df39db8d5f585dc540be41b61a650) for the original context.

We want to turn the interprocedural fixpoint into a proper functor.
This will properly encapsulate the global fixpoint state, preventing outside mutations.

Unfortunately, the missing flow analysis (which turns unknown calls into sinks) currently writes to the internal fixpoint state to add models, in the middle of fixpoint iterations. This works fine because those models do not require new propagations, but this is in general unsafe and we do not want to allow this in the future. It also adds implicit call graph edges during fixpoint iterations, which is also fine in that case because they do not need to be in the dependency graph, but is unsafe in general.

This diff makes the following changes:
* Explicitly add edges to the symbolic "unknown callee" object during call graph construction;
* Explicitly generate models for "unknown callee" object before the fixpoint;
This way, the missing flow analysis does not require to update the state of the global fixpoint or add call graph edges during fixpoint iterations.

Reviewed By: dkgi

Differential Revision: D35991464

fbshipit-source-id: 6fd0600b2a62aa65d2e93748d9eaac12ef29c51e