Do not store test models directly in the fixpoint state
Summary:
See the context in D35833510 (https://github.com/facebook/pyre-check/commit/a6b4b31c5896af7849ae4f216178d7a8035c5288).

In future diffs, we will disallow to write directly to the fixpoint state.
Unfortunately, this is what we do in a test helper function.
To help this refactoring, let's make the test helper return the set of initial models, instead of storing it in the fixpoint state. The test is then responsible for calling the appropriate APIs.

Reviewed By: stroxler

Differential Revision: D36017956

fbshipit-source-id: 050483771e941ffdc97cb49d95f578299542bb59