Use soft tabls
Summary:
We already get tabs rather than spaces from codemirror's auto-indent,
but hitting the TAB key inserts hard tabs.

This can be a major issue if hard and soft tabs get mixed, because this
will produce a python syntax error - and an error with a not great
message at that!

There's actually no purely declarative way to do soft tabs in
codemirror (at least version 5), the recommended way straight from
the docs is a key remapping, which is what we do here.

Reviewed By: grievejia

Differential Revision: D36045104

fbshipit-source-id: b842ac6eb2e0e6d0d55b36b462a0c7b403b229fa