Use `pre` to format errors
Summary:
Most of our errors to date are one-liners, which means that
using a div with monospace formatting works okay.

But the new `reveal_locals` implementation uses line breaks,
and these can't be correctly displayed without a `pre` tag;
using `pre` has a side benefit that the default css leads
to some light grey background that makes it really easy to see
which lines belong to the same error.

Reviewed By: alexkassil

Differential Revision: D36044995

fbshipit-source-id: 18d18d1ac6361a9b4272106b704bd0ef236632a9