Update the criteria for deciding whether two classes have a descendant relation
Summary:
For the known problem in the previous diff, the root cause is, multi-inheritance breaks the assumption that, the intersection of two class intervals is empty iff. one is the descendant of the other.

A key observation is that, under multi-inheritance, the intersection of any two class intervals A and B is
- either empty,
- or A (which means A is a subclass of B),
- or B (which means B is a subclass of A),
- or non-empty but neither A nor B (which means there is no descendant relation between A and B. The intersection is non-empty only because A and B share children).

The correctness of the above statement is based on the assumption that, class A is a descendant of class B iff. class A's interval is fully contained in class B's interval. We have guaranteed this when computing class intervals.

Reviewed By: arthaud

Differential Revision: D35725213

fbshipit-source-id: 3902b489a4bec7bbbd48b8d7669c0b1530451119