Use IntervalSet instead of ClassInterval everywhere
Summary:
Now that we have implemented a set representation of class intervals, we can use it during the analysis.

# Known Problem

Note, the false positives in the integration tests that are caused by multi-inheritance are not removed by this diff, which is surprising but does not indicate an implementation bug. Instead, there is a bug in the design. For example, the result of intersecting B with C (see below) is non-empty, which violates the assumption that, the intersection of two class intervals is non-empty iff. they have a descendant relation. The result of B intersecting C is [3,4], because B and C share a same child.
```
           A: [1,8]
          / \
         /   \
        /     \
B: [2,5]   C: [6,7]; [3,4]
        \     /
         \   /
          \ /
           D: [3,4]
```
The next diffs will fix this problem.

Reviewed By: arthaud

Differential Revision: D35690212

fbshipit-source-id: 5f5df460e22668f526a398083d244d84333bf7f7