Reorder AstEnvironment module
Summary:
I'm trying to get some trivial reorderings out of the way before working
toward first-class overlayable tables in the enviornment stack. We need
the reordering because each `t` type (the mutatable environment itself)
is going to need to include a table, and as a result the module defining
the shared memory table needs to come first, above the `t`.

Keeping the reordering (which is free of logic changes) in its own diff
will make later changes much easier to review.

Reviewed By: dkgi

Differential Revision: D36172143

fbshipit-source-id: ba01bcf46571c678729f6c045edca511fb5e56e1