Reorder TypeEnvironment code
Summary:
In order to make environments overlayable, I'm going to need each `t` type (that defines
a mutable environment) to include a handle on a first-class shared memory table.

It follows that I need the shared memory modules to be defined above the `t` types so that
they can be used. Doing this reordering ahead of time in logic-free diffs will make the
overlayable environment logic easier to review.

Reviewed By: dkgi

Differential Revision: D36172152

fbshipit-source-id: 2c3448fee8d1e2ee7be0469cab2d71c06b5ac437