s/Function/Define in UGE
Summary:
Consistent names make a big difference to code readability; we were
inconsistently mixing 'function' and 'define body' in the naming of the
define keys.

Instead, let's just use 'define' everywhere. It's better than 'define body'
because the key is for a define, and isn't specific to "body" (and all
else equal shorter names > longer names); it's better than 'function' because
it's more clear that methods and module/class toplevels are included.

Reviewed By: pradeep90

Differential Revision: D36173339

fbshipit-source-id: c0f86d10ffee86f753f8f9b8ef28bd5698eae323