Disable caching of the call graph
Summary:
Unfortunately, caching the call graph is unsound since it can be invalidated. Indeed, adding a source or sink on an attribute will change the call graph (add edges), hence we need to recompute it entirely.

In the future, we could potentially cache the call graph and only discard it if new attribute models were introduced.

Differential Revision: D36267161

fbshipit-source-id: 9cb9e4dcd249d87eff3eda1a943fb9882d7edaa9