Reveal a call graph construction bug when property setters are defined in functions, not classes
Summary:
It turns out that property setters can be defined in functions, in addition to being defined in classes.
Hence, when property setters are defined in functions, we consider them to be a same target as the properties themselves, thus causing they to share a same call graph -- the analysis stores a mapping from targets to call graphs.

Reviewed By: arthaud

Differential Revision: D36209754

fbshipit-source-id: 3b083bb616b973e1efd42d79a7ca38fd2a7f5ed9