Make go-to-def work for methods and attributes.
Summary:
Problem: We were unable to resolve  the location for method calls, such as `foo.bar()`. That is because we were trying two things: (a) the whole reference as such, in case it matched a global function or something or (b) resolving the base and then trying to get the location of `<base>.attribute`.

However, (b) failed completely for go-to-def because we don't store methods in the global location lookup table.

Solution: Resolve the type of base, get the class summary, and lookup the attribute location.

Reviewed By: grievejia

Differential Revision: D36284230

fbshipit-source-id: b8c77e1b04c231521570829313c659a8f3a25def