Turn the interprocedural analysis into a functor
Summary:
# Context
The `pyre analyze` command was first meant to be a generic command that would call into different static analyses (e.g, dead stores, taint analysis, etc). This is implemented with the `Interprodecural` framework that runs a global fixpoint, given a callgraph. Different analyses can be plugged in by calling `Interprocedural.AnalysisResult.Make (ResultType).Register (AnalysisType)`.

Unfortunately, with time, the `pyre analyze` command has been changed specifically for the taint analysis, to a point where even the entrypoint calls taint analysis specific functions (inlining decorators, computing class intervals, etc.). This is introducing technical debt, for instance, with parameters being threaded in more and more complicated ways.

The call graph construction of the interprocedural framework is also heavily specialized for the taint analysis (e.g, it adds taint-it-taint-out flags). This is  to a point where it is unsound for other analyses, which was discovered when working on a type inference analysis.

# Proposed changes
We reached a point where we want to change the whole design of the interprocedural framework.

The proposed diff turns the registration-based interprocedural framework into an OCaml functor (module `Interprocedural.FixpointAnalysis` with a `Make` submodule), which is generic and can be instantiated multiple times for different analysis.

This also introduces a `cleanup` function which removes the fixpoint state from shared memory. This is useful to prevent integration tests having side effects on other tests.

# Upsides
This diff removes a lot of the GADT usages and `Obj.magic` calls, which were pretty hard to read and maintain.
The global fixpoint state is now hidden in the implementation, which means it is not exposed anymore. This prevents subtle bugs that introduce non determinism.
This also introduces a proper `Registry` module, which is a mapping from callable to models. This avoids a lot of duplication in our codebase (e.g, merging two mappings is frequent).

# Trade-offs
The current interprocedural framework allows to run multiple global fixpoints in parallel (for instance, a dead store and a taint analysis). We now lose that ability with the proposed changes. In practice, this was never used.

Another downside would potentially be code duplication - most likely, for the entrypoint. I think this can be avoided with proper designs.

Reviewed By: dkgi

Differential Revision: D36336305

fbshipit-source-id: fce26fca38b6c298705403bdb0b30e57f8bf9c4e