Remove UnannotatedGlobalEnvironment.WriteOnly
Summary:
Before I do any more big refactors of environments working toward support
for overlays and unsaved changes, I'm going to do a few small fixes to fix
nonstandard aspects of UGE that make it harder to read and edit than it
could be.

Here, I remove the WriteOnly module. This module is
- nonstandard (other environments, including AstEnvironment, store their
  read/write data in module scope and the readonly view in ReadOnly, this
  approach works fine and leads to less indentation
- not really helpful as an abstraction - the tiny bit of function hiding
  we get isn't meaningfully contributing to modularity, and the extra
  `sig` section is just boilerplate making it harder to skim our code
- leading to code duplication because the module-level `read_only`
  redirects to the `WriteOnly.read_only`; this isn't terrible, but it
  makes code navigation slightly harder for no actual benefit

Removing it is a small win but a win nontheless, and will be a bigger
win soon when I need to make environment tables first-class; the actual
motivation for this change is that I started working on that and realized
I was getting too much duplication between WriteOnly and module scope.

Reviewed By: grievejia

Differential Revision: D36180998

fbshipit-source-id: e185d61bfa8708466d47b09e4f0e259342bfa8dd