Move key clearing logic into KeyTracker
Summary:
When I started working on using first-class tables for key tracker,
which is something we need to do for overlays, I realized how messy
it is for us to be doing a series of tightly coupled reads and
mutations against KeyTracker shared-memory tables directly in the
update logic.

Let's push that code down to the level of a single well-defined
api operation inside of KeyTracker.

The resulting struct is a little ugly - with redundant list and set
forms of each previous key output - because unfortunatley we're
tightly coupled to specific types and we need both forms (lists for
dependency tracking and sets for the update result).

Reviewed By: grievejia

Differential Revision: D36186354

fbshipit-source-id: c8f59fc82cc4d4d7f49727449b17874b9108619e