Overlayable environments: Make KeyTracker first-class
Summary:
**Context and what this diff does at a high level**

In terms of the amount of fiddly shared-mem state management, the
UnannotatedGlobalEnvironment is the hardest one to change because
it has so many parts: each of the key tracker tables, plus each
of the dependency tracked tables is handled separately.

As a result, I'm working on gradually creating a more first-class-style
api where operations have a handle to the read/write environment as
a value. I'm starting by making the KeyTracker tables first class. We'll
eventually need to make tham overlayable rather than just first-class,
but the necessary code changes for first-class memory are a subset
so it makes sense to switch to first-class first.

**Ugliness: the base environment needs to reuse the same table in all tests**

An ugly thing I realized is that the way we run our tests requires us to
reuse the same key tracker for all base environments. The root cause
is that shared memory tables can't be properly garbage collected, which
causes two problems:
1. temporarily, while the dependency-tracked tables are first-class,
    failing to reuse the key tracker causes us to not delete data, which
   catastrophically fails the tests
2. even once we have first-class tables, it might be a bad idea to use them
    in base environments because this would lead to a memory leak; instead
    of repeatedly clearing out and reusing the same shared memory hashmaps,
   we would create new data for every test which could eat up RAM in long
   test suites.

Note that both these problems are fundamentally issues with shared memory;
like a *lot* of our ugly environment api gotchas they will go away if we can
ever move to multicore and normal ocaml tables :)

**Details**

Note that once the environment is properly first-class, the lazy loader
needs to take both a read/write `UnannotatedGlobalEnvironment.t` and
an `AstEnvironment.ReadOnly.t` as input because it needs readonly access
to upstream AST changes, but the ability to mutate the UGE as it lazy
loads. Currently I pass them as separate arguments which is verbose; I'll
clean that up in a follow-up.

Because my ultimate goal is to make all of the tables first-class
rather than just the key tracker tables, I pass the entire environment
to any function that mutates one of the dependency tracked tables; given
the current state of the code I could have just passed the key tracker
but that would cause thrash in subsequent commits.

Reviewed By: grievejia

Differential Revision: D36183757

fbshipit-source-id: 8cb384c7450543e0564dcf6c43d540089e8d6f74