Overlays: use first-class modules for dependency tracked tables
Summary:
This commit is the most sweeping single-commit API change needed to make
our environments overlayable: it creates first-class values for all of the
dependency tracked tables that back our environments.

When combined with the first-class KeyTracker change from a previous commit,
this means the entire environment is first class. Dependencies are still
global but that's not a problem from the perspective of overlays; having extra
dependencies is a perf hit (a tiny one in practice since unsaved changes
dependencies will tend to get saved at some point anyway) but cannot affect
correctness.

As with the first-class key tracker, currently `create` always reuses the
same base table values; this is an ugly hack but is needed to avoid leaking
memory in long tests suites, because (unlike in production) we create many
base environments, and using separate first-class values would lead to the
shared memory not being reused. We could look into removing this in a follow
up if the RAM use isn't too high; from a correctness point of view it should
be okay as long as the leaking memory isn't a problem.

None of the tables are overlayable yet - I still need to roll out overlay
logic (which will implement the first-class table api) but this lets us
get all the plumbing in place.

Reviewed By: grievejia

Differential Revision: D36206455

fbshipit-source-id: a8fc7289edc8e60c7b6a556e979c66df80475ded