Lift t / create / etc to the top of UnannotatedGlobalEnvironment
Summary:
The definition of an environment's read-write type and
related helper functions ought to go right below the definition
of the shared memory data, but above the initialization and
update logic.

This organization makes the types easier to see; it will also be
mandated by ocaml's ordering restrictions as soon as we make
tables first-class because the `t` type will have to be defined
above all functions that try to mutate a `t`.

Reviewed By: grievejia

Differential Revision: D36178673

fbshipit-source-id: e1028e0600c000d87a4d7f7247ca3e7cbda99fc9