Tighten up lazy loading and read-only apis
Summary:
In incurred some tech debt in the run up to first-class key tracker
(D36183757) because I kept passing around the environment and ast
environment directly to LazyLoading instead of defining a type
to have a nicely-encapsulated abstraction. The same problem spilled
over to the `ReadOnlyTable.S` interface.

Here I fix the problem by:
- renaming LazyLoading to LazyLoader to emphasize that it is becoming
  a first-class thing
- defining LazyLoader.t, and making lazy load operations take this as
  input
- defining a ReadOnlyTable.S.t, which is currently implemented as
  just a LazyLoader.t but is abstract because as soon as we flip to
  first-class dependency-tracked-tables we'll need to also pass a handle
  to the table itself

The resulting API is considerably more modular, and should be easier
to adapt to first-class dependency-tracked-tables which are coming up
soon.

Reviewed By: grievejia

Differential Revision: D36186355

fbshipit-source-id: c36607d1d8cbedcd6ddd1cf2af2d97e1a1157ac9