Use consistent ordering in UnannotatedGlobalEnvironment
Summary:
Something that kept mildly bothering me when I was working on the lazy
parser was that UnannotatedGlobalEnvironment has quite a lot of different
tables in it, but the order in which we use those tables varies between
different blocks of code.

Here I reorganize functions and tables so that similar operations live
together and we always go in the same low-to-high-cardinality ordering:
- modules
- then classes
- then defines
- then globals

I also snuck in a rename `s/get_define_body_keys/get_define_keys/g` that
I missed in an earlier renaming commit; it's not worth trying to restack.

Reviewed By: grievejia

Differential Revision: D36180916

fbshipit-source-id: 665e01f9f1e65cda12be88a035649891220349f7