Fix a couple more out-of-order operations
Summary:
For a module like UnannotatedGlobalEnvironment that wraps a lot
of data all sitting at the same level, a key part of readability is
to always do things in a consistent order when the order doesn't matter
so that it's easier to skim and see when a block of code is "just do the
same thing to every table".

I tried to convert everything to use the ordering
modules -> class_summaries -> function_definitions -> unannotated_globals
earlier but I keep seeing places I missed; here I update the transaction
and direct_data_purge logic to use the standard ordering.

Reviewed By: grievejia

Differential Revision: D36207572

fbshipit-source-id: c8cd5fb29d0bef2d4ffc43084143ec30e0ebef89