Split find references for local and global names
Summary:
We probably can't support find all references with a table lookup, because we need to be able to handle requests on any name - not just function calls, and not just at the point of definition.

For example,

```
1: def foo() -> None:
2:    x = 1
3:    y = x
4:
5: def bar() -> None:
6:    foo()
7:
8: def baz() -> None:
9:    foo()
```

In this case, we'd need to be able to ask for "find all references" on `foo` on line 9, and return the `foo`s on lines 1 and 6. We need to be able to ask for "find all references" on `x` in line 2 and return the `x` on line 3.

Essentially, if we represented this as a mapping from reference -> list of references, every single reference in the list would also need its own entry to refer back to the rest of the group; so this isn't just a reverse callgraph lookup.

Given that we also need to support this for essentially every valid name: functions, attributes, globals, local values, class names, module names, etc., I think the only way to implement this is to take Hack's approach of tweaking the name of the reference in question, and then re-asking the type checker to calculate undefined name errors and converting those errors into reference locations.

In cases where we're asking for references of a name that is in the global scope, this starts to look a lot like unsaved changes overlays with a few exceptions (cc stroxler):

- Unsaved changes can be filtered down to only a module in question to prevent fan-out and keep things performant, also limit the memory footprint of writing overlay keys to all of the dependency tables. For find-all-references, we /need/ the fan-out and don't mind the perf hit, but this means we will need to make adjustments to make sure we aren't blowing up memory if we do many find-all-references checks in a row on the same server.
- To clean up the memory footprint, we'll either need to pay a heavy performance cost of remembering the last overlaid change and reverting it when applying the new overlaid change (I don't think this makes sense, since perf costs are highly variable depending on the name we pass in), or implementing a clean-up/sweep functionality that goes through shared memory tables and uses the dependency graph to delete keys written by the overlay check when we are done.
- We'll want to block writes to the dependency graph and just rely on the "logical" key to propagate our changes and propagate the cleanup.
- We'll need a transform visitor to figure out the right place to make the AST renaming change to kick off the incremental check.

In cases where we're asking for references of a name that is only in a local scope, we don't have to care about the environment at all - just take the relevant define body, make an adjustment to make sure the line defining the reference doesn't register or is modified, and then run an uninitialized local analysis on that one define and convert the errors to reference locations.

I'll tackle the local portion first before diving into the overlay/environment dependency changes needed to support this globally.

Note that: performance is likely not a concern here at all. find-all-references in Hack will load for minutes if you invoke it on some standard library function or basic class that is used all over the repository, as it should. Find all references should behave, performance wise, identically to a type check when the thing is changed. Also fun fact that pylance's find all references is also very slow and anecdotally seems quite flaky/often incorrect. I could not get a find all references request working at all in the current jedi setup.

Reviewed By: stroxler

Differential Revision: D35629810

fbshipit-source-id: 4761013b68a1389287731f9b9ac36fc749defcf3