Create test cases to describe the expected analysis results of __setitem__
Summary:
As described in T100215741, we do not track TITO information precisely when analyzing `dict.__setitem__` (i.e., taint a single key, instead of the entire dictionary), which leads to a false positive. The reason is we have an imprecise model for `dict.__setitem__` (which is used in the production. See https://fburl.com/code/cypv4br9):
```
def dict.__setitem__(self, __key: TaintInTaintOut[Updates[self]], __value: TaintInTaintOut[Updates[self]]): ...
```
Note that, without the above model, the current analysis is able to track the TITO information precisely.

To precisely (but not necessarily correctly) analyze  `dict.__setitem__` and to correctly (but not necessarily precisely) analyze `sub_class.__setitem__` for any `sub_class` that overrides `__setitem__`, the solution is to:
- Remove the above imprecise model.
- When analyzing `e.__setitem__`, we case split based on the type of expression `e`:
  - When `e` is of type `dict`, we use a built-in model for `__setitem__` (see more explanation below). Although such treatment is incorrect (because `e` can be actually a strict subtype of `dict`, whose `__setitem__` has a different behavior than `dict.__setitem__`), we consider this as a way to improve the SNR.
  - When `e` is not of type `dict`, we use either the user-provided or the inferred model for `__setitem__`.
- The built-in model for `dict.__setitem__` is to let assignment `e[key] = value` to taint the `key` field of expression `e` (as opposed to the entire expression).

This diff creates test cases whose analysis results will eventually (i.e., not now) describe how we intend to analyze `__setitem__`. The next diffs in this stack will realize such intent.

Reviewed By: arthaud

Differential Revision: D36299096

fbshipit-source-id: 828f6768dc77f3352eaf2917b171ce2d24ce0c26